requirements/dev.txt
-r live.txt
django-extensions
django-nose
six
nose
